requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
boto==2.48.0
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
colorclass==2.2.0
Cython==0.26
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-auth0==0.0.3
django-autocomplete-light==3.2.9
django-celery==3.2.1
django-celery-beat==1.0.1
django-compat==1.0.14
django-cors-headers==2.1.0
django-freshdesk==1.0.1
django-hijack==2.1.4
django-hijack-admin==2.1.4
django-storages==1.6.5
django-tastypie==0.14.0
djrill==2.1.0
docopt==0.6.2
enum34==1.1.6
feedparser==5.2.1
gunicorn==19.7.1
html5lib==1.0b8
idna==2.6
ipaddress==1.0.18
kombu==3.0.37
Logentries==0.17
lxml==3.8.0
olefile==0.44
packaging==16.8
Pillow==4.2.1
pip-upgrader==1.4.3
psycopg2==2.7.3
pyasn1==0.3.2
pycparser==2.18
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.1
python-freshdesk==1.0.0
python-mimeparse==1.6.0
pytz==2017.2
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
setupfiles==0.0.50
six==1.10.0
South==1.0.2
sqlparse==0.2.3
terminaltables==3.1.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
whitenoise==3.3.0
xhtml2pdf==0.0.6
